requirements.txt
# Core NLP libraries
underthesea==6.7.0
pyvi==0.1.1
transformers==4.36.0
datasets==2.15.0

# Deep learning
torch==2.1.0
tensorflow==2.14.0  # optional, for some models

# Data processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utils
tqdm==4.66.1
pyyaml==6.0.1
jupyter==1.0.0
ipywidgets==8.1.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0